requirements.txt
Pillow
tensorflow>=2.0.0
numpy
datetime
pathlib